Move follow resolver to user
